+++ src/powerquery-language-services/analysis/analysisOptions.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { LibrarySymbolProvider, SymbolProvider } from "./providers";
+import { LibrarySymbolProvider, SymbolProvider } from "../providers";
 
 export interface AnalysisOptions {
     readonly locale?: string;
